requirements.txt
asgiref==3.5.2
astroid==2.11.7
async-generator==1.10
atomicwrites==1.4.1
attrs==21.4.0
beautifulsoup4==4.11.1
boto==2.49.0
boto3==1.24.36
botocore==1.27.36
CacheControl==0.12.11
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
dill==0.3.5.1
dj-database-url==1.0.0
Django==4.0.6
django-addanother==2.1.0
django-anymail==8.6
django-autocomplete-light==3.9.4
django-client-side-image-cropping==0.1.10
django-cors-headers==3.13.0
django-ses==3.1.0
django-storages==1.13
djangorestframework==3.13.1
firebase==3.0.1
firebase-admin==5.2.0
git-filter-repo==2.34.0
google==3.0.0
google-api-core==2.8.2
google-api-python-client==2.54.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.6.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpcio==1.47.0
grpcio-status==1.47.0
gunicorn==20.1.0
h11==0.13.0
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.7.1
lxml==4.9.1
MarkupSafe==2.1.1
mccabe==0.7.0
msgpack==1.0.4
outcome==1.2.0
packaging==21.3
Pillow==9.2.0
platformdirs==2.5.2
pluggy==1.0.0
proto-plus==1.20.6
protobuf==3.20.1
psycopg2==2.9.3
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pylint==2.14.5
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-decouple==3.6
pytils==0.4.1
pytz==2022.1
requests==2.28.1
rsa==4.8
s3transfer==0.6.0
selenium==4.3.0
six==1.16.0
smtp2go==2.0.0
smtp2go-django==2.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
tomlkit==0.11.4
tools==0.1.9
trio==0.21.0
trio-websocket==0.9.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.10
Werkzeug==2.1.2
whitenoise==6.2.0
wrapt==1.14.1
wsproto==1.1.0
